requirements.txt
certifi
chardet
DBUtils
idna
PyMySQL
redis
requests
urllib3
xlwings
selenium
Pillow
baidu-aip
alibabacloud_ocr_api20210707
alibabacloud_darabonba_stream
tencentcloud-sdk-python